1. add python_magic to requirements.txt for verifying pdf filetype in
   `metadata_scan.py`, saving the user the effort of removing remnants
of failed downloads.
2. added a bunch of logic to `metadata_scan.py` to this end as well.
  a) compare filenames of pdf files with titles of metadata lookups to
ensure the correct information was pulled from the api lookup. added a
fuzzy matching function with a threshold value so that the researcher
can bypass manual verification for clear matches.
  b) added propmt option to open pdf directly from the script for manual
verification or serach for isbn/doi info, saving the researcher from
needing to locate the file in the filesystem and open manually open it.
As a bonus result. It seems to load much faster than it would that way.
  c) verified remove option functioning correctly.
  d) updated prompt format so that difference between matched or not
matched is much more salient...

I started these changes not expecting much, but with the sample data I
used for testing/debugging I'd estimate a 3-5x reduction in time to
complete the manual parts of the research gathering process. All sets
are different though, but I'm happy :).